backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
celery>=5.3.6
redis>=5.0.1
biopython>=1.83
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.3
torch>=2.2.0
motor>=3.3.0
beanie>=1.25.0
python-dotenv>=1.0.1
pydantic-settings>=2.0.0
email-validator>=2.1.0
python-multipart>=0.0.9
requests>=2.31.0
jinja2>=3.1.3
rdkit>=2023.9.5
httpx>=0.26.0
groq>=0.4.0
